move tbc to standalone class
