+++ packages/commonwealth/server/util/tokenBalanceCache/tokenBalanceCache.ts
@@ -1,66 +1,19 @@
 import { RedisCache } from 'common-common/src/redisCache';
-import { DB } from '../models';
-import { BalanceSourceType } from '../util/requirementsModule/requirementsTypes';
-import { __getCosmosNativeBalances } from './server_token_balance_methods/get_cosmos_balances';
-import { __getErc1155Balances } from './server_token_balance_methods/get_erc1155_balances';
-import { __getErc20Balances } from './server_token_balance_methods/get_erc20_balances';
-import { __getErc721Balances } from './server_token_balance_methods/get_erc721_balances';
-import { __getEthBalances } from './server_token_balance_methods/get_eth_balances';
+import { DB } from '../../models';
+import { BalanceSourceType } from '../requirementsModule/requirementsTypes';
+import { __getCosmosNativeBalances } from './providers/get_cosmos_balances';
+import { __getErc1155Balances } from './providers/get_erc1155_balances';
+import { __getErc20Balances } from './providers/get_erc20_balances';
+import { __getErc721Balances } from './providers/get_erc721_balances';
+import { __getEthBalances } from './providers/get_eth_balances';
+import {
+  Balances,
+  GetBalancesOptions,
+  GetCosmosBalancesOptions,
+  GetEvmBalancesOptions,
+} from './types';
 
-type GetCosmosBalanceOptions = {
-  balanceSourceType: BalanceSourceType.CosmosNative;
-  addresses: string[];
-  sourceOptions: {
-    cosmosChainId: string;
-  };
-};
-
-type GetEvmBalancesBase = {
-  addresses: string[];
-  sourceOptions: {
-    evmChainId: number;
-  };
-};
-
-type GetErc20BalanceOptions = GetEvmBalancesBase & {
-  balanceSourceType: BalanceSourceType.ERC20;
-  sourceOptions: {
-    contractAddress: string;
-  };
-};
-
-type GetErc721BalanceOptions = GetEvmBalancesBase & {
-  balanceSourceType: BalanceSourceType.ERC721;
-  sourceOptions: {
-    contractAddress: string;
-  };
-};
-
-type GetErc1155BalanceOptions = GetEvmBalancesBase & {
-  balanceSourceType: BalanceSourceType.ERC1155;
-  sourceOptions: {
-    contractAddress: string;
-    tokenId: string;
-  };
-};
-
-type GetEthNativeBalanceOptions = GetEvmBalancesBase & {
-  balanceSourceType: BalanceSourceType.ETHNative;
-};
-
-type GetEvmBalancesOptions =
-  | GetEthNativeBalanceOptions
-  | GetErc20BalanceOptions
-  | GetErc721BalanceOptions
-  | GetErc1155BalanceOptions;
-
-export type GetBalancesOptions =
-  | GetEvmBalancesOptions
-  | GetCosmosBalanceOptions;
-
-export type Balances = { [address: string]: string };
-
-export class ServerTokenBalanceController {
+export class TokenBalanceCache {
   constructor(public models: DB, public redis: RedisCache) {}
 
   /**
@@ -85,7 +38,7 @@ export class ServerTokenBalanceController {
     return balances;
   }
 
-  private async getCosmosBalances(options: GetCosmosBalanceOptions) {
+  private async getCosmosBalances(options: GetCosmosBalancesOptions) {
     const chainNode = await this.models.ChainNode.scope(
       'withPrivateData',
     ).findOne({
